elevator-pipeline/poll_outage/requirements.txt
alembic==0.9.6
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
dataset==1.0.2
idna==2.6
lxml==4.1.0
Mako==1.0.7
MarkupSafe==1.0
normality==0.4.4
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
